backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
openai>=1.12.0
python-multipart>=0.0.9
httpx>=0.27.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
httpx>=0.27.0
